refactor: extract toTeamDrawResult utility and improve type definitions
- Move toTeamDrawResult function from EditTeamsPage to new edit-teams utils file
- Rename TeamHeader component to TeamHeaderRow for clarity
- Consolidate DragListItem type definitions using BaseItem intersection types
- Update TeamHeaderRow to use DragListItem type instead of any
- Add null coalescing for playerCount prop
- Remove commented-out bench-header rendering code
- Reorder DragListItem union type members
